+++ app/Filament/Resources/HorizonDeploymentResource/Pages/EditHorizonDeployment.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\HorizonDeploymentResource\Pages;
+
+use App\Filament\Resources\HorizonDeploymentResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditHorizonDeployment extends EditRecord
+{
+    protected static string $resource = HorizonDeploymentResource::class;
+}