+++ app/Filament/Resources/HorizonDeploymentResource/Pages/ListHorizonDeployments.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\HorizonDeploymentResource\Pages;
+
+use App\Filament\Resources\HorizonDeploymentResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListHorizonDeployments extends ListRecords
+{
+    protected static string $resource = HorizonDeploymentResource::class;
+}